удалила моки
